fix workday and test lottery

+++ src/index.ts
@@ -6,7 +6,7 @@ import { ConfigService } from './config.service'
 import { StorageService } from './storage.service'
 import { ACTION_TEXT } from './constants'
 import { Action, SlackUser, PostEvent } from './models'
-import { separateData } from './utils'
+import { lottery, separateData } from './utils'
 
 declare var global: any
 
@@ -31,8 +31,12 @@ function send_message(data: object) {
 }
 
 global.send_question = (event): void => {
-  ConfigService.remove_timer(event.triggerUid)
+  console.info('send_question')
+  if (event) {
+    ConfigService.remove_timer(event.triggerUid)
+  }
   let today = new Date()
+  console.info(today)
   StorageService.prepareStorage(today)
   let actions: Action[] = []
   let times = ConfigService.get_times()
@@ -106,7 +110,9 @@ global.draw = (event): void => {
       message += ` ----- ${time}\n`
       teams.forEach(function(team: SlackUser[]) {
         team_no += 1
-        if (team.length > 1 && lottery(lottery_ratio)) message += '当たり！'
+        if (lottery(team, lottery_ratio)) {
+          message += '当たり！'
+        }
         message += ` チーム ${team_no}: `
         team.forEach(function(person) {
           message += `<@${person.userId}> ,`
@@ -124,11 +130,4 @@ global.set_timer = (): void => {
 }
 global.init = (): void => {
   ConfigService.initialize()
-}
-
-function lottery(ratio: number): boolean {
-  if (ratio > 0) {
-    return 1 == Math.ceil(Math.random() * ratio)
-  }
-  return false
 }
